kpno_allsky/requirements.txt
numpy>=1.15.4
scipy>=1.0.0
matplotlib>=2.2.2
astropy>=2.0.3
pyephem>=3.7.6.0
requests>=2.20.1
pandas>=0.22.0
pillow>=4.2.1
pytesseract>=0.2.6 
python-daemon>=2.2.3